[addons] change binary interface (webbrowser required parts)

+++ xbmc/addons/kodi-addon-dev-kit/include/kodi/gui/dialogs/Select.h
@@ -107,7 +107,7 @@ namespace dialogs
     inline int Show(const std::string& heading, const std::vector<std::string>& entries, int selected = -1, unsigned int autoclose = 0)
     {
       using namespace ::kodi::addon;
-      unsigned int size = entries.size();
+      unsigned int size = static_cast<unsigned int>(entries.size());
       const char** cEntries = (const char**)malloc(size*sizeof(const char**));
       for (unsigned int i = 0; i < size; ++i)
       {
@@ -162,7 +162,7 @@ namespace dialogs
     inline int Show(const std::string& heading, std::vector<SSelectionEntry>& entries, int selected = -1, unsigned int autoclose = 0)
     {
       using namespace ::kodi::addon;
-      unsigned int size = entries.size();
+      unsigned int size = static_cast<unsigned int>(entries.size());
       const char** cEntries = static_cast<const char**>(malloc(size*sizeof(const char*)));
       for (unsigned int i = 0; i < size; ++i)
       {
@@ -227,7 +227,7 @@ namespace dialogs
     inline bool ShowMultiSelect(const std::string& heading, std::vector<SSelectionEntry>& entries, int autoclose = 0)
     {
       using namespace ::kodi::addon;
-      unsigned int size = entries.size();
+      unsigned int size = static_cast<unsigned int>(entries.size());
       const char** cEntryIDs = static_cast<const char**>(malloc(size*sizeof(const char*)));
       const char** cEntryNames = static_cast<const char**>(malloc(size*sizeof(const char*)));
       bool* cEntriesSelected = static_cast<bool*>(malloc(size*sizeof(bool)));
